Fixed load_all_sessions()
at line 262ish we used to set a struct session equal to the value session that we want and then use that session variable when parsing instead of the original meaning that we were parsing to a copy of the session not the actual session. I fixed it by removing this copy and making it parse to the original.